requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pythonping==1.1.4
PyYAML==6.0.2
requests==2.32.3
serverchan-sdk==1.0.6
setuptools==75.1.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
wakeonlan==3.1.0
Werkzeug==3.1.3
wheel==0.44.0
